Change API to have Immutable Spans
This one is a game changer.  Here, we assert that the implementation of
Money will be immutable.  This has several effects:

1. We remove the concept of a timeout all together.  There are no span
managers that keep track of spans.  It is up to the thread (and maybe
some support not yet written) to manage spans.  This is very
interesting, as it forces you to ask the question about why we needed
timeouts in the first place.

2. Threadlocal gets a little muddied.  I tinkered with it a little in
this PR, but a workable threadlocal solution would likely involve a
wrapper Span that managed that instead of how it is in the
DefaultSpanFactory.

3. We needed to change the definition of the start/stop Timer.  Since
these change the internal state of the Span, startTimer had to return a
Span instead of a Timer.

-----
The biggest change is the removal of the timeout.  It is really
interesting, but in an immutable model I would have to understand why we
need timeouts, I may just be missing something.